+++ packages/observable-client/src/chainHead/storage-queries.ts
@@ -17,7 +17,8 @@ export const getRecoveralStorage$ = (
     isHighPriority: boolean,
   ): Observable<StorageItemResponse> =>
     new Observable<StorageItemResponse[] | Observable<StorageItemResponse>>(
-      (observer) =>
+      (observer) => {
+        let nDiscarded = 0
         getFollower().storageSubscription(
           hash,
           queries,
@@ -29,12 +30,6 @@ export const getRecoveralStorage$ = (
             observer.error(error)
           },
           () => {
-            observer.complete()
-          },
-          (nDiscarded) => {
-            // TODO: leave it like this b/c due to a bug on
-            // PolkadotSDK sometimes this value is `undefined`
-            // https://github.com/paritytech/polkadot-sdk/issues/6683
             if (nDiscarded > 0)
               observer.next(
                 recoveralStorage$(
@@ -44,8 +39,16 @@ export const getRecoveralStorage$ = (
                   true,
                 ),
               )
+            observer.complete()
           },
-        ),
+          (_nDiscarded) => {
+            // TODO: leave it like this b/c due to a bug on
+            // PolkadotSDK sometimes this value is `undefined`
+            // https://github.com/paritytech/polkadot-sdk/issues/6683
+            if (_nDiscarded > 0) nDiscarded = _nDiscarded
+          },
+        )
+      },
     ).pipe(mergeAll(), withRecovery(isHighPriority))
 
   return recoveralStorage$